Fix lli, this is a hack!
The uart baud tate of 115200 is desired to achieve higher sample
rate of the IMU, currently it is observed that the LLI cannot
handæle communicatio form the HLI when using 115200, so for now we
are using 57600, which works, but we have slow IMU sampling. Should
be good enough so far.