Controller working kind of... but not as we want it
It uses the hacky llisim topic to send the thrustdiff number. This is
not what we need, we need it to use the LLIinpyt message to send the
control inputs. This does currently not work at all. The boat is just
jittering onear its position. Maybe the LLIinput data field cannot
contain enough resolution for the thing. But this is strange.The LLI
uses a 16 bit int for the PWM. It should be ok with +- 500.
Above/under this should be in saturating. 300 is very high force,
around 50 N.